Splitting up medUtilities
